+++ src/js/theme.js
@@ -1,0 +1,11 @@
+const theme = {
+  primary: '#00e5ff',
+  secondary: '#00acc1',
+  accent: '#00695c',
+  error: '#D50000',
+  warning: '#FFFF00',
+  info: '#00B8D4',
+  success: '#76FF03'
+}
+
+export default theme