+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 # Changelog
 
-## 2.0
+* [2.0.0-alpha](#2.0.0-alpha)
+
+---
+
+## 2.0.0-alpha
+
+---
+
+### 2.0.0-alpha2
+
+* Webpack, VueJS and Vuetify support added
+* Theme with new colors
+* New architecture
+* Players selection output in the console
+* Updated README.md with documentation
+
+### 2.0.0-alpha1
 
 * Updated README.md
 * rewriting from zero